+++ DCMods_Main/Common_Videos.cpp
@@ -132,7 +132,7 @@ void DisplayVideoFadeout(int fadeout, int mode)
 	Direct3D_SetZFunc(7u);
 	Direct3D_EnableZWrite(0);
 	if (mode == 0) njDrawTriangle2D_SomeOtherVersion((NJS_POINT2COL*)&VideoFadeoutFrame, 4, -1000.0f, NJD_TRANSPARENT | NJD_FILL);
-	else DrawRect_Queue(0.0f, 0.0f, HorizontalResolution, VerticalResolution, 22048.0f, fadeout << 24, QueuedModelFlagsB_EnableZWrite);
+	else DrawRect_Queue(0.0f, 0.0f, (float)HorizontalResolution, (float)VerticalResolution, 22048.0f, fadeout << 24, QueuedModelFlagsB_EnableZWrite);
 	Direct3D_EnableZWrite(1);
 	Direct3D_ResetZFunc();
 }
@@ -303,7 +303,7 @@ void OnInitEnd_Videos()
 {
 	if (DLLLoaded_HDVideos)
 	{
-		for (int i = 0; i < LengthOfArray(VideoDataArray); i++)
+		for (unsigned int i = 0; i < LengthOfArray(VideoDataArray); i++)
 		{
 			VideoDataArray[i].Width = Videos[0].Width;
 			VideoDataArray[i].Height = Videos[0].Height;
